test(reddog): split lease security matrix
Worker-Lane: reddog-authority-runtime

Slice: REDDOG_EXTERNAL_SIGNER_AUTHORITATIVE_USE_LEASE_PHASE1
