+++ modules/communication/moltbot_bridge/ModLog.md
@@ -3,13 +3,17 @@
 - Added a strict typed exact-effect request and opaque one-shot process capability; the signer recomputes effect identity, rejects noncanonical JSON, and binds current manifest, generation, owner config, run packet, session, socket, profile, and key.
 - Extended the existing socket client v2 grant envelope and durable E0 replay store. Lease-domain socket v1 rejects, and client rehydration has no injected verifier, replay, time, or parent-authority verdict; the existing E0 exact-request grant remains the only signing authority.
 - Root current-generation resolution now selects the exact signer profile, public key, and key epoch from the authenticated config. The signed request binds all four E0 replay identities, both signer and client enforce the same store, lease expiry cannot exceed selection expiry, and monotonic process-local deadlines prevent clock-rollback revival.
-- Added a composed socket-v2 -> durable E0 grant -> resolve-per-sign Ed25519 signer -> external issuer -> real WRE effect regression, plus local-key, split-root, generation-overrun, and rollback attacks.
+- Added a composed socket-v2 -> durable E0 grant -> resolve-per-sign Ed25519 signer -> external issuer -> real WRE effect regression, plus local-key, split-root, generation-overrun, and rollback attacks; split contract and adversarial cases into bounded test modules.
 - Bound real leases through both worktree admission registries and the direct WRE spine using the exact executor-plan and valve-decision digests.
 - No issuer, resolver activation, worker, repository, PR, merge, HoloIndex, or live-canary effect was added; other effects and crash recovery remain blocked. (WSP 00/15/22/50/62/71/97)
 ## 2026-08-11: Architect and verifier contract reconciliation
-- Replaced retired `AUDIT` with canonical `AUDIT_NO_EFFECT`; read-only `FIX` determinations remain evidence and signed plans carry exact test scope.
-- Forwarded exact work order, commit, policy, and test scope into the independent producer/verifier without authority expansion.
-- Local differential evidence remains diagnostic; autonomous admission requires external verifier authority and reservation. (WSP 00/15/22/50/62/97)
+- Replaced the retired literal `AUDIT` bootstrap default with the canonical
+  `AUDIT_NO_EFFECT` progressive-stage constant. Read-only architect `FIX`
+  determinations remain audit evidence; signed plans can carry exact test scope.
+- Forwarded exact work order, commit, signed policy, and resolved test scope into
+  the independent producer and verifier without changing execution authority.
+- Local differential evidence is diagnostic only; autonomous admission still
+  requires an external verifier authority and reservation. (WSP 00/15/22/50/62/97)
 ## 2026-08-11: Bounded iterative repository grounding
 - Extended the existing transport-neutral grounding service with at most two
   deterministic HoloIndex owner queries per semantic target under one global
