+++ modules/communication/moltbot_bridge/tests/test_reddog_wsp62_security_repair_exemptions.py
@@ -57,13 +57,11 @@ EXPECTED_MODULE_FILES = {
 }
 PROVIDER_EVIDENCE_EXACT_FILES = {
     "INTERFACE.md",
-    "ModLog.md",
     "src/reddog_backend_architect_determination_runtime.py",
     "src/reddog_fusion_progress_validation.py",
     "src/reddog_provider_call_evidence.py",
     "src/reddog_readonly_0102_audit_worker_runtime.py",
     "src/reddog_readonly_audit_task_executor.py",
-    "tests/TestModLog.md",
     "tests/test_reddog_backend_architect_determination_runtime.py",
     "tests/test_reddog_fusion_progress_receipt.py",
     "tests/test_reddog_main_readonly_operational_bootstrap.py",
@@ -73,6 +71,7 @@ PROVIDER_EVIDENCE_EXACT_FILES = {
     "tests/test_reddog_resident_architect_durable_agentdb_cycle.py",
     "wsp_62_exemptions.yaml",
 }
+AUDIT_LOG_FILES = {"ModLog.md", "tests/TestModLog.md"}
 PUBLICATION_MODULE_FILES = {
     "src/reddog_architect_fix_promotion_publication.py",
     "src/reddog_architect_fix_promotion_publication_validation.py",
@@ -236,6 +235,26 @@ def test_provider_evidence_exemptions_match_exact_touched_file_sizes() -> None:
             assert item["no_growth_ceiling"].get("functions", {}) == (
                 _oversized_function_sizes(target)
             )
+
+
+def test_required_audit_logs_use_non_blocking_archival_policy() -> None:
+    entries = {
+        item["file"]: item
+        for item in yaml.safe_load(
+            (MODULE_ROOT / "wsp_62_exemptions.yaml").read_text(encoding="utf-8")
+        )["exemptions"]
+    }
+    advisory = {
+        path: item
+        for path, item in entries.items()
+        if item.get("enforcement_mode") == "advisory_archive"
+    }
+    assert set(advisory) == AUDIT_LOG_FILES
+    for item in advisory.values():
+        assert item["advisory_archive_threshold"] == 1000
+        assert item["owner"] and item["architect_reviewer"]
+        assert item["remediation"]
+        assert "no_growth_ceiling" not in item
 
 
 def test_architect_publication_modules_need_no_wsp62_exemption() -> None:
